Solve issue with regex for finding tags by attribute hitting the backtrack limit on large strings
The previous way of finding tags by attribute was very prone to causing the PREG_BACKTRACK_LIMIT_ERROR to occur when running the regex.

I wasn't able to find a way to optimise the regex while still keeping it functional to reduce the backtracking, so instead I've re-implemented the way we
find tags by attribute.

The new approach should look directly for the actual attribute(s) definition, and then proceed to work backwards/forwards to get the full tag string.

The two approaches to tag finding (by tag or by attribute) are now in separate classes with a common interface, since they now work so differently.