Remove '+' from character class
This was a typo as it was supposed to be the quantifier for the character class. Since the previous capturing group will capture greedily, this should only ever
match a single character anyway, so the quantifier can just be removed.

We could make the previous group lazy, but that will cause catastrophic backtracking, and we're not really that bothered about perfectly capturing the tag name.